Split compliance test into ONPRC and WNPRC variants
